ai-service/requirements.txt
fastapi
uvicorn
crewai
google-genai
python-dotenv
requests
pillow
faiss-cpu
sentence-transformers
httpx